% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{geocode}
\alias{geocode}
\title{Get geocode from address string.}
\usage{
geocode(address)
}
\arguments{
\item{address}{Character vector of address strings.}
}
\description{
Get geocode from address string.
}
\examples{
address <- "東京都中央区月島２－１０－１"
geocode(address)

}
