% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{has_na}
\alias{has_na}
\title{Does each column in a data.frame have NA}
\usage{
has_na(df)
}
\arguments{
\item{df}{data.frame}
}
\description{
Does each column in a data.frame have NA
}
\examples{
df <- data.frame(a=1:3, b=5:7, c=c(NA, 1:2))
has_na(df)

}
