% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{is_numeric}
\alias{is_numeric}
\title{A more general test of an object being interpretable as numbers.}
\usage{
is_numeric(x, na = c("", "NA"))
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{na}{Character vector of strings to use for missing values.
Set this option to \code{character()} to indicate no missing values.}
}
\description{
A more general test of an object being interpretable as numbers.
}
\examples{
x <- c(1, 2, NA, "AAA", "123")
is_numeric(x)

}
