% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mode}
\alias{mode}
\title{Mode Value}
\usage{
mode(x)
}
\arguments{
\item{x}{vector(numeric or character)}
}
\description{
Compute the sample mode value
}
\examples{
mode(c(1,2,3,3,3,6))
mode(c("A", "A", "A", "B"))

}
