% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scientific_10}
\alias{scientific_10}
\title{Get 10^x format expression for ggplot2 label}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Get 10^x format expression for ggplot2 label
}
\examples{
df <- data.frame(x=10^(-2:1),y=10^(-2:1))
ggplot(df, aes(x, y)) +
  geom_point() +
  scale_x_continuous(label=scientific_format()) +
  scale_y_continuous(label=scientific_format())

}
