% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{is_numeric}
\alias{is_numeric}
\title{A more general test of an object being interpretable as numbers.}
\usage{
is_numeric(x, na = FALSE)
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{na}{Boolean. It determines wether NA is regarded as TRUE/FALSE.}
}
\description{
A more general test of an object being interpretable as numbers.
}
\examples{
x <- c(1, 2, NA, "AAA", "123")
is_numeric(x)

}
